Move d3 edge logic to edge utils file
